Afford a twilio SMS Authenticator for keycloak
This project creates a `lib-all.jar` that can be placed in keycloak's
directory named `/providers`. When present, the jar affords an SMS
authenticator. The Authenticator implementation is based on dasniko's
at https://github.com/dasniko/keycloak-2fa-sms-authenticator and also
requires his keycloak-requiredaction jar to be in `/providers`. After
the jars are in place, keycloak needs to be configured to use them.
See https://www.n-k.de/2020/12/keycloak-2fa-sms-authentication.html
and
https://github.com/PhilanthropyDataCommons/auth/issues/2#issuecomment-1440797906
for more details.

The twilio settings are loaded via environment variables:
 - `TWILIO_PHONE_NUMBER`: the "from" phone number set up in twilio.
 - `TWILIO_ACCOUNT_SID`: the SID or username for twilio API access.
 - `TWILIO_AUTH_TOKEN`: the token or secret for twilio API access.

The LICENSE continues the license of the original work by Niko Köbler.
See
https://github.com/dasniko/keycloak-2fa-sms-authenticator/blob/4205a6/LICENSE

The jar is a shaded fat jar, but avoids including classes found on the
keycloak classpath.

Issue #2 Two-factor SMS authentication with keycloak and twilio

Co-authored-by: Jason Owen <jason@jasonaowen.net>