Serve login page fonts from the auth server
This change prevents cross-origin resource sharing issues, out-of-date
URL issues, and privacy issues. The fonts for the login page are
served by the same server that serves the page, namely the Keycloak
server.

These fonts came from the PDC pilot service on 2023-04-05 from URLs
found in the previous commit.

While we cringe at putting binaries in the repository, we would like
to use this jar imminently. There is no intermediate step in the
upstream `data-viewer` project that produces and publishes binaries,
but that may be a useful longer-term goal: pull the best binaries
as part of a build here (i.e. `../gradlew jar`) from the binaries
produced in the upstream `data-viewer` build/publish pipeline.

Issue #7 Theme the user-facing Keycloak pages