Include fonts from webjars (not source control)
Declare a dependency on a webjar for Source Sans Pro font. The jar
includes the font files we need. Use a gradle build task to extract
the files from the jar into the resources directory at build time.
The `woff` and `woff2` files get included in the resulting theme jar.
This way we are not committing binary files to source control.

Issue #7 Theme the user-facing Keycloak pages